perf(gas): batch 4a — degenerette unrolls, curve reorder, coinflip/whale/time caching
Council-verified + adversarial diff-review SHIP (8/8) gas wins:
- degenerette-05/06: unroll fixed-4 _score / _countGoldQuadrants (byte-identical
  transcriptions; shifts 0/8/16/24 and 3/11/19/27)
- degenerette-07: unchecked the per-spin payout accumulators (totalPayout +
  acc.ethClaimable/flipMint/wwxrpMint); the once-per-bet survival-flip
  subtraction stays checked (underflow guard preserved)
- libs-04: reorder the activity-curve cap branch below the two segment checks
  (bit-identical; knees 235/300/305 < 500 < 30000, disjoint partition)
- libs-07: widen currentDayIndexAt param uint48->uint256, drop the timestamp cast
- coinflipflip-02: thread PlayerCoinflipState pointer into _claimCoinflipsInternal
  across all 9 callers (incl. the sDGNRS/redeemableFlipBacking sites)
- whalefoil-02: reuse cached passLevel at the 70/30-vs-95/5 pool split
- whalefoil-03: thread passLevel-1 into _rewardDeityPassDgnrs

Gate GREEN 1225/0/103; adversarial diff-review 8/8 SHIP (0 fix, 0 revert).

+++ contracts/libraries/GameTimeLib.sol
@@ -19,7 +19,7 @@ library GameTimeLib {
      * @return Current day index (1-indexed from deploy day).
      */
     function currentDayIndex() internal view returns (uint24) {
-        return currentDayIndexAt(uint48(block.timestamp));
+        return currentDayIndexAt(block.timestamp);
     }
 
     /**
@@ -29,7 +29,7 @@ library GameTimeLib {
      *        earlier values underflow the unsigned subtraction and wrap.
      * @return Day index (1-indexed from deploy day).
      */
-    function currentDayIndexAt(uint48 ts) internal pure returns (uint24) {
+    function currentDayIndexAt(uint256 ts) internal pure returns (uint24) {
         uint24 currentDayBoundary = uint24((ts - JACKPOT_RESET_TIME) / 1 days);
         return currentDayBoundary - uint24(ContractAddresses.DEPLOY_DAY_BOUNDARY) + 1;
     }
